+++ src/cluster/DotNext.Net.Cluster/Net/Multiplexing/ProtocolCommand.cs
@@ -7,7 +7,7 @@ internal abstract class ProtocolCommand
     public abstract void Write(IBufferWriter<byte> buffer);
 }
 
-internal sealed class StreamRejectedCommand(ulong streamId) : ProtocolCommand
+internal sealed class StreamRejectedCommand(uint streamId) : ProtocolCommand
 {
     public override void Write(IBufferWriter<byte> buffer) => Protocol.WriteStreamRejected(buffer, streamId);
 }
